Initial commit - add zephyr QPc port
This commit adds a minimal port of QPc for Zephyr. It's implementation
is based on POSIX/pthread interface which is provided by Zephyr.

The existing POSIX/pthread port of QPc could not be used as-is, because
there are some slight differences in POSIX interface between Zephyr and
glibc on Linux. Besides, by reimplementing it for Zephyr, APACHE license
could be used, which makes this port more suitable for inclusion in
Zephyr codebase.